backend/ml_service/requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.15.0
sentence-transformers
numpy
scikit-learn
pydantic
tf-keras
prometheus-fastapi-instrumentator
aio-pika